+++ src/test/java/edu/harvard/iq/dataverse/api/UtilIT.java
@@ -3686,6 +3686,19 @@ public class UtilIT {
         return response;
     }
 
+    static Response clearDatasetCitationDateField(String datasetIdOrPersistentId, String apiToken) {
+        String idInPath = datasetIdOrPersistentId; // Assume it's a number.
+        String optionalQueryParam = ""; // If idOrPersistentId is a number we'll just put it in the path.
+        if (!NumberUtils.isCreatable(datasetIdOrPersistentId)) {
+            idInPath = ":persistentId";
+            optionalQueryParam = "?persistentId=" + datasetIdOrPersistentId;
+        }
+        Response response = given()
+                .header(API_TOKEN_HTTP_HEADER, apiToken)
+                .delete("/api/datasets/" + idInPath + "/citationdate" + optionalQueryParam);
+        return response;
+    }
+
     static Response getFileCitation(Integer fileId, String datasetVersion, String apiToken) {
         Boolean includeDeaccessioned = null;
         return getFileCitation(fileId, datasetVersion, includeDeaccessioned, apiToken);