+++ doc/sphinx-guides/source/api/changelog.rst
@@ -13,7 +13,7 @@ v6.4
 - For the ``oai_dc`` metadata export and harvesting (OAI-PMH) format:
 
   - dc:type was mapped to the field "Kind of Data". Now it is hard-coded to the word "Dataset".
-  - dc:date was mapped to the field "Production Date" when available and otherwise to "Publication Date". Now it is mapped only to the field "Publication Date".
+  - dc:date was mapped to the field "Production Date" when available and otherwise to "Publication Date". Now it is mapped the field "Publication Date" or the field used for the citation date, if set (see :ref:`set-citation-date-field`).
   - dc:rights was not mapped to anything. Now it is mapped (when available) to terms of use, restrictions, and license.
 
 v6.3